+++ client/my-sites/subscribers/components/subscriber-launchpad/subscriber-launchpad.tsx
@@ -37,6 +37,7 @@ const SubscriberLaunchpad = ( {
 
 			<Launchpad
 				siteSlug={ site?.slug ?? null }
+				site={ site }
 				checklistSlug={ checklistSlug }
 				onPostFilterTasks={ taskFilter }
 				launchpadContext={ launchpadContext }